+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Fix filter pattern to use `and` instead of `or` in filtering so the event is _allowed_ if it is allowed by **both** spot and the filterValue (_unless one of those allowed lists is empty_)
+- Add trace logs to filter pattern
+- Add `LoggerFactory.common` function to help create a common logger factory for kafka application usage
 
 ## 15.2.0 - 2021-10-29
 - Update dependencies